pyproject.toml
[project]
name = "aig-response-tracing"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click==8.1.8",
    "openai==1.63.2",
    "pangea-sdk==5.5.1",
]

[tool.uv]
dev-dependencies = [
    "mypy==1.15.0",
    "ruff==0.9.6",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # ruff defaults
    "E4",
    "E7",
    "E9",
    "F",

    "I",   # isort
    "PTH", # flake8-use-pathlib
]
